Move net::resource_task unit tests into the unit_tests crate.
