+++ tests/unit/net/mod.rs
@@ -1,3 +1,4 @@
 mod cookie;
 mod data_loader;
 mod image_cache_task;
+mod resource_task;
